requirements.txt
certifi==2017.4.17
chardet==3.0.4
cycler==0.10.0
Django==1.11.2
django-extensions==1.7.9
django-material==1.0.0
idna==2.5
matplotlib==2.0.2
numpy==1.13.0
oauthlib==2.0.2
olefile==0.44
pandas==0.20.2
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
urllib3==1.21.1
